Fixed Cross Origin Conflicts :lock:
Cross origin problems are the bane of the people who don't use local
servers.

Before you take your pitchfork, all the ones I know are over engineered
for the solution I need. Besides, it doesn't always works.

I really hope a stray developer sees this and replace his local server
with a script similar to this.

By the way, that is a Windows Batch file.